backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
langchain
transformers
pypdf
python-multipart
PyPDF2